refactor: extract shared watcher session helpers and rename events-parser
Eliminates ~55 lines of duplicated session-update logic between ClaudeJsonlWatcher
and CopilotCliDetector by extracting applyActivityUpdate, applyModelUpdate, and
applySummaryUpdate into watcher-session-helpers.ts. Renames events-parser.ts to
copilot-cli-jsonl-parser.ts to match the naming convention of its Claude counterpart.
