+++ backend/src/services/copilot-cli-jsonl-parser.ts
